# src/routes-build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CreateTrickDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"videoURL": {
						"type": "string"
					},
					"level": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"videoURL",
					"level"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrickDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"videoURL": {
						"type": "string"
					},
					"level": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"videoURL",
					"level"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTrickDto": {
				"$ref": "#/components/schemas/CreateTrickDto"
			},
			"UserAuthenticationResponseDto": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAuthenticationRequestDto": {
				"properties": {
					"facebookAccessToken": {
						"type": "string"
					}
				},
				"required": [
					"facebookAccessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenResponseDto": {
				"$ref": "#/components/schemas/UserAuthenticationResponseDto"
			},
			"UserRole": {
				"enum": [
					"ADMIN",
					"CUSTOMER",
					"REFEREE",
					"INSTRUCTOR"
				],
				"type": "string"
			},
			"ChangeUserRoleRequestDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "wake-server-side",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/tricks": {
			"post": {
				"operationId": "CreateTrick",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTrickDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTricks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrickDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/tricks/{id}": {
			"delete": {
				"operationId": "DeleteTrickById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTrickById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrickDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTrickDto"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserAuthenticationResponseDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserAuthenticationRequestDto"
							}
						}
					}
				}
			}
		},
		"/users/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RefreshTokenResponseDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/users/change-role": {
			"patch": {
				"operationId": "ChangeRole",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"JWT": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangeUserRoleRequestDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}